Refactored some Startup functions to their own class
